Refactor plot functions
